+++ jcm/physics/clouds/speedy_condensation.py
