+++ jcm/physics/speedy/surface_flux.py
@@ -9,7 +9,7 @@ from jcm.physics.speedy.params import Parameters
 from jcm.physics_interface import PhysicsTendency, PhysicsState
 from jcm.physics.speedy.physics_data import PhysicsData
 from jcm.physics.speedy.physical_constants import p0, rgas, cp, alhc, sbc, grav
-from jcm.physics.speedy.humidity import get_qsat, rel_hum_to_spec_hum
+from jcm.physics.clouds.speedy_humidity import get_qsat, rel_hum_to_spec_hum
 from jcm.utils import pass_fn
 
 @jit
